Move server to module
There's no functional difference here but it will make the server code
easier to test since it will be importable in a normal way.
